Namespace the media services under Spree::Media
MediaDeletion and MediaUsage become Spree::Media::Destroy and
Spree::Media::Usage, matching how every other service names itself after the
model it acts on (Carts::Destroy, Orders::Fees::Destroy).

The SDK's MediaUsageReference type keeps its name — it is published API and
the rename buys nothing there.
