+++ docs/plans/6.0-media-library.md
@@ -89,7 +89,7 @@ the storage level.
   from "refuse until the merchant clears each use"). `DELETE /admin/media/:id`
   on a file in use returns 422 with the usage list; passing `detach=true` —
   which the dashboard sends after a confirmation naming the count — runs
-  `Spree::MediaDeletion`, removing the file from every placement and plain
+  `Spree::Media::Destroy`, removing the file from every placement and plain
   attachment in its store, then deleting the row so the blob purges. Clearing
   each use by hand was busywork the usage list already knew how to do; the
   refusal survives as the default so an API client cannot pull a file out from
@@ -185,7 +185,7 @@ the storage level.
 Blob-level usage lookup (for the delete flow and the media detail panel): given
 a row, find all `active_storage_attachments` sharing its blob ids, resolve to
 owner records, and present them grouped. This is a read-side service —
-`Spree::MediaUsage` — not columns.
+`Spree::Media::Usage` — not columns.
 
 ### Admin API
 
@@ -247,7 +247,7 @@ grid reuses the existing small/thumbnail renditions.
 
 - Never strip or "align" the `active_storage_attachments` → blobs foreign key
   with the no-FK convention — blob-sharing deletion safety depends on it.
-- `Spree::MediaUsage` must stay store-scoped on every one of its three sources
+- `Spree::Media::Usage` must stay store-scoped on every one of its three sources
   (attachments, placements, rich text) — a blob can be shared, but a file's
   reported uses are its own store's rows.
 - Rich text search covers content models only (`CONTENT_MODELS`); never add
